tests: move task_description test to its own TrossenMCAPBackend suite
